moved cli.py to scripts, removed command line tests, added basic parser
