+++ src/main.py
@@ -1,0 +1,4 @@
+from scripts.cli import CLI
+
+cli_program = CLI("command_line_interface")
+cli_program.start()
